+++ packages/editor/editor-confluence-transformer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/editor-cq-transformer
 
+## 2.10.1
+
+## 2.10.0
+- [minor] Add full width and wide layout support for single image [ae72acf](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ae72acf)
+
+## 2.9.10
+
 ## 2.9.9
 
 ## 2.9.8