+++ packages/editor/editor-json-transformer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/editor-json-transformer
 
+## 2.5.15
+
+## 2.5.14
+
 ## 2.5.13
 
 ## 2.5.12