+++ packages/editor/editor-markdown-transformer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/editor-markdown-transformer
 
+## 0.2.8
+
+## 0.2.7
+
 ## 0.2.6
 
 ## 0.2.5