+++ packages/editor/editor-wikimarkup-transformer/CHANGELOG.md
@@ -1,3 +1,7 @@
 # @atlaskit/editor-wikimarkup-transformer
 
+## 0.0.3
+
+## 0.0.2
+
 ## 0.0.1